Fix resize transform
Resize() takes a sequence of values like (720, 1280) when there are multiple. They need to be enclosed in brackets. Since each value was passed as a separate parameter, Resize() read the second value as the interpolation mode. The values were also listed in the wrong order. Additionally, the InterpolationMode needed to be specified as NEAREST to maintain logical validity of seg maps